requirements.txt
Flask
gunicorn
requests
google-auth
google-analytics-data
pytrends
pandas